+++ SmartReceipts/WBReceiptsHelper.h
@@ -53,7 +53,7 @@
 -(WBReceipt*) copyReceipt:(WBReceipt*) receipt fromTrip:(WBTrip*)oldTrip toTrip:(WBTrip*) newTrip;
 -(WBReceipt*) moveReceipt:(WBReceipt*) receipt fromTrip:(WBTrip*)oldTrip toTrip:(WBTrip*) newTrip;
 -(BOOL) deleteWithParent:(NSString*) parent inDatabase:(FMDatabase*) database;
--(BOOL) deleteWithId:(int) receiptId forTrip:(WBTrip*) currentTrip;
+-(BOOL) deleteWithId:(NSUInteger) receiptId forTrip:(WBTrip*) currentTrip;
 -(BOOL) swapReceipt:(WBReceipt*) receipt1 andReceipt:(WBReceipt*) receipt2;
 
 -(NSString*) selectCurrencyForReceiptsWithParent:(NSString*) parent inDatabase:(FMDatabase*) database;
